VideoGenerationCode/PythonFlask/requirements.txt
Flask==2.1.2
pymongo[srv]==4.1.1 
ffmpeg-python==0.2.0
requests==2.27.1
python-dotenv==0.20.0
cloudinary==1.29.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.45.0
werkzeug==2.0.3
moviepy==1.0.3
google-generativeai==0.1.0

